pyproject.toml
[tool.poetry]
name = "poetry-sample-project"
version = "0.1.0"
description = ""
authors = ["ytakayama"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.2"
pgvector = "^0.2.5"
psycopg2-binary = "^2.9.9"
streamlit = "^1.31.0"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.6.0"
langchain = "^0.1.6"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.2"
pytest = "^7.3.2"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"